fix(renderer): fixed some rendering bugs (#47)
* fix(renderer): fixed some rendering bugs

Fixed various rendering bugs. First, a problem is solved where the value was always passed as string
in the dom, which was incorrect. Secondly, if a single value is passed in the dom, the value is not
passed anymore as string. Lastly, there was a issue with resubscribing listeners to emits, while
this was not needed.

* fix(unit test): fix internal renderer util test

* feat(validator): implement nested field validation